Extract AbstractFindStructureRequest
